requirements.txt
black==21.9b0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
gitdb==4.0.7
GitPython==3.1.24
idna==3.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
regex==2021.9.30
requests==2.26.0
smmap==4.0.0
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
websocket-client==1.2.1
